docs: Improved comments
