+++ milestone-3/frontend/src/App.tsx
@@ -7,11 +7,14 @@ import { BrowserRouter, Route, Routes } from "react-router-dom";
 
 function App() {
   return (
+    // BrowswerRouter to control the website navigation
     <BrowserRouter>
+    {/* Always render Navbar component */}
       <Navbar />
       <Routes>
         <Route path="/" element={<Home />} />
         <Route path="/about" element={<About />} />
+        {/* This path will render the relevant RecipePage based on the id parameter given */}
         <Route path="/recipe/:id" element={<RecipePage />} />
       </Routes>
     </BrowserRouter>
