+++ milestone-3/frontend/src/components/About.tsx
@@ -1,3 +1,5 @@
+//About page
+
 export default function About() {
   return (
     <main>
