+++ milestone-3/frontend/src/components/Navbar.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import { Link } from "react-router-dom";
 import "./Navbar.css";
 
+//Navbar that is rendered on each page. Uses react router links.
 export default function Navigation() {
   return (
     <header className="header">
