+++ milestone-3/frontend/src/recipeData.ts
@@ -1,6 +1,18 @@
-import {RecipePreviewData} from "./components/RecipePreview"
+/**
+ * Data for the recipes is stored in this file.
+ */
 
-export interface Recipe extends RecipePreviewData{
+//Data object for Recipe preview cards
+export interface RecipePreviewData {
+  id: number;
+  name: string;
+  description: string;
+  image: string;
+  imgAlt: string;
+}
+
+//Using inheritance since both interfaces share properties
+export interface Recipe extends RecipePreviewData {
   ingredients: string[];
   instructions: string[];
 }
